+++ lib/controllers/layout_controller.dart
@@ -1,3 +1,4 @@
+import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 
 import '../controllers/transaction_controller.dart';
@@ -39,7 +40,7 @@ class LayoutController extends ChangeNotifier with WidgetsBindingObserver {
     try {
       final wallet = _walletController.getWalletById(transaction.walletId ?? 1);
       await _transactionController.createTransaction(transaction);
-      BigInt deductedFromReserved = wallet.autoExpanse(
+      Decimal deductedFromReserved = wallet.autoExpanse(
         transaction.amount,
         useReserved: useReserved,
       );
@@ -47,9 +48,9 @@ class LayoutController extends ChangeNotifier with WidgetsBindingObserver {
       await _walletController.updateWallet(wallet);
       await _walletController.loadData();
 
-      BigInt deductedFromRegular = transaction.amount - deductedFromReserved;
+      Decimal deductedFromRegular = transaction.amount - deductedFromReserved;
 
-      if (deductedFromRegular > BigInt.zero && !excludeDaily) {
+      if (deductedFromRegular > Decimal.zero && !excludeDaily) {
         _userController.useDaily(deductedFromRegular);
       }
