+++ lib/controllers/profile_controller.dart
@@ -1,3 +1,4 @@
+import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 
 import 'user_controller.dart';
@@ -19,7 +20,7 @@ class ProfileController with ChangeNotifier {
     }
   }
 
-  Future<bool> saveBaseDailyLimit(BigInt amount) async {
+  Future<bool> saveBaseDailyLimit(Decimal amount) async {
     try {
       _userController.updateBaseDailyLimit(amount);
       await _userController.saveUser();
@@ -30,7 +31,7 @@ class ProfileController with ChangeNotifier {
     }
   }
 
-  Future<bool> saveEmergencyAmount(BigInt amount) async {
+  Future<bool> saveEmergencyAmount(Decimal amount) async {
     try {
       _userController.updateEmergencyAmount(amount);
       await _userController.saveUser();
